Crate from_back

Library for indexing and slicing from the back of a sequence.

The main entrypoint for this library is the idx! macro, which generates the library’s structs. These structs can be used to index into common sequence types, notably Vec and slice. The idx! macro exposes a virtual unary operator ^, which indicates the index should count from the back of the sequence instead of the front.

Examples

let vec = vec![8, 6, 7, 5, 3, 0, 9];
// the element second from the back is 0
assert_eq!(vec[idx!(^2)], 0);
// slice the elements two from the front to three from the back (exclusive)
assert_eq!(&vec[idx!(2..^3)], &[7, 5]);
// slice the elements two from the front to three from the back (inclusive)
assert_eq!(&vec[idx!(2..=^3)], &[7, 5, 3]);

Macros

idx

Create an index or range which may have “from back” components.

Structs

SeqRange

A parallel to std::ops::Range (idx!(start..end))

SeqRangeFrom

A parallel to std::ops::RangeFrom (idx!(start..))

SeqRangeInclusive

A parallel to std::ops::RangeInclusive (idx!(start..=end))

Enums

SeqIndex

Container type for indexing from the front or back of a sequence (idx!(index))